requirements.txt
cycler==0.10.0
decorator==4.4.1
imageio==2.8.0
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.1.3
networkx==2.4
numpy==1.18.1
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sk-video==1.1.10
sklearn==0.0#
torch==1.4.0
torchvision==0.5.0#
